+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/user/dto/request/v3/UserInfoRequestDto.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.domain.user.dto.request;
+package com.hertz.hertz_be.domain.user.dto.request.v3;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.hertz.hertz_be.domain.user.entity.enums.AgeGroup;
